Misc spec update for calico/node daemonset
- disable conntrack INVALID rule through env, which conflicts with DSR
- ignore loose RPF mode (if set) through env
- enable rollingupdate strategy
- turn-off masquerade for pre-configured IP pool

By default IP masquerading is ON for the IP pool defined in the same
spec (192.168.0.0/16). However, for VIP address not in this range,
there's an IPTable rule generated by Felix doing source NAT and replace
it with the Node IP which is running traffic originating Pod. This is
not desired behavior when destination IP is a VIP that doesn't belong to
Calico IPAM pool. Before it is fixed at the model level, applying this
patch as a quick workaround.